+++ cmd/gravinet/main.go
@@ -46,7 +46,7 @@ import (
 
 // Build metadata, overridable via -ldflags.
 var (
-	version = "455"
+	version = "456"
 	commit  = "none"
 )